% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleClassroom.R
\name{createGoogleCourse}
\alias{createGoogleCourse}
\title{Create a Course Section}
\usage{
createGoogleCourse(name, section, ownerId)
}
\arguments{
\item{name}{The name of the course}

\item{section}{The name of the course section}

\item{ownerId}{The owner of the created course who is added as a teacher}
}
\value{
Information on the newly created course.
}
\description{
This function creates a Google Classroom Course
}
\section{References}{

\href{https://developers.google.com/classroom/reference/rest/v1/courses/create}{Google API Documentation}
}

\concept{Classroom}
