% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdminDirectory.R
\name{getGoogleCustomerID}
\alias{getGoogleCustomerID}
\title{Get a Google User's Customer ID.}
\usage{
getGoogleCustomerID(customerKey)
}
\arguments{
\item{customerKey}{Id of the customer to be retrieved}
}
\value{
The customer key of the Google user
}
\description{
This function returns the customer key for a Google user.
}
\section{References}{

\href{https://developers.google.com/admin-sdk/directory/v1/reference/customers/get}{Google API Documentation}
}

\concept{Users}
