% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdminDirectory.R
\name{getGoogleGroupSettings}
\alias{getGoogleGroupSettings}
\title{Get a Google group's settings}
\usage{
getGoogleGroupSettings(groupUniqueId)
}
\arguments{
\item{groupUniqueId}{The group's email address.}
}
\value{
The Google group's settings
}
\description{
This function retrieves a group's settings identified by the group email address.
}
\section{References}{

\href{https://developers.google.com/admin-sdk/groups-settings/v1/reference/groups/get}{Google API Documentation}
}

\concept{Groups}
