% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdminDirectory.R
\name{getGoogleUser}
\alias{getGoogleUser}
\title{Get a Google User.}
\usage{
getGoogleUser(
  userKey,
  projection = NULL,
  customFieldMask = NULL,
  viewType = NULL
)
}
\arguments{
\item{userKey}{Identifies the user in the API request. The value can be the user's primary email address, alias email address, or unique user ID.}

\item{projection}{What subset of fields to fetch for this user. Acceptable values are: "basic": Do not include any custom fields for the user (default), "custom": Include custom fields from schemas requested in customFieldMask, and "full": Include all fields associated with this user.}

\item{customFieldMask}{A comma-separated list of schema names. All fields from these schemas are fetched. This should only be set when projection=custom.}

\item{viewType}{Whether to fetch the administrator-only or domain-wide public view of the user. Acceptable values are "admin_view": Results include both administrator-only and domain-public fields for the user. (default), and "domain_public": Results only include fields for the user that are publicly visible to other users in the domain.}
}
\value{
Information on the Google user.
}
\description{
This function returns information on a Google user.
}
\section{References}{

\href{https://developers.google.com/admin-sdk/directory/v1/reference/users/get}{Google API Documentation}
}

\concept{Users}
