% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoogleClassroom.R
\name{listGoogleCourses}
\alias{listGoogleCourses}
\title{List Google Courses}
\usage{
listGoogleCourses(studentId = NULL, teacherId = NULL, courseStates = NULL)
}
\arguments{
\item{studentId}{Restricts returned courses to those having a student with the specified identifier. The identifier can be one of the following: the numeric identifier for the user, the email address of the user, the string literal "me", or indicating the requesting user}

\item{teacherId}{Restricts returned courses to those having a teacher with the specified identifier. The identifier can be one of the following: the numeric identifier for the user, the email address of the user, the string literal "me", or the indicating the requesting user}

\item{courseStates}{Restricts returned courses to those in one of the specified states The default value is ACTIVE, ARCHIVED, PROVISIONED, DECLINED.}
}
\value{
A list of Google Classroom courses.
}
\description{
This function lists all Google Classroom Courses by student, teacher or domain
}
\section{References}{

\href{https://developers.google.com/classroom/reference/rest/v1/courses/list}{Google API Documentation}
}

\concept{Classroom}
