% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdminDirectory.R
\name{listGoogleUsersInDomain}
\alias{listGoogleUsersInDomain}
\title{List Google Users in a Domain}
\usage{
listGoogleUsersInDomain(
  domain,
  orderBy = NULL,
  sortOrder = NULL,
  showDeleted = NULL,
  query = NULL,
  projection = NULL,
  customFieldMask = NULL,
  viewType = NULL,
  customer = NULL
)
}
\arguments{
\item{domain}{The domain name. Use this field to get fields from only one domain. To return all domains for a customer account, use the customer query parameter instead. Either the customer or the domain parameter must be provided.}

\item{orderBy}{Property to use for sorting results. Acceptable values are: "email": Primary email of the user, "familyName": User's family name, and "givenName": User's given name.}

\item{sortOrder}{Whether to return results in ascending or descending order. Acceptable values are: "ASCENDING": Ascending order, and "DESCENDING": Descending order.}

\item{showDeleted}{If set to true, retrieves the list of deleted users. (Default: false)}

\item{query}{Query string for searching user fields. For more information on constructing user queries, see {https://developers.google.com/admin-sdk/directory/v1/guides/search-users} {Search for Users}.}

\item{projection}{What subset of fields to fetch for this user. Acceptable values are: "basic": Do not include any custom fields for the user (default), "custom": Include custom fields from schemas requested in customFieldMask, and "full": Include all fields associated with this user.}

\item{customFieldMask}{A comma-separated list of schema names. All fields from these schemas are fetched. This should only be set when projection=custom.}

\item{viewType}{Whether to fetch the administrator-only or domain-wide public view of the user. Acceptable values are "admin_view": Results include both administrator-only and domain-public fields for the user. (default), and "domain_public": Results only include fields for the user that are publicly visible to other users in the domain.}

\item{customer}{The unique ID for the customer's G Suite account. In case of a multi-domain account, to fetch all groups for a customer, fill this field instead of domain. You can also use the my_customer alias to represent your account's customerId. The customerId is also returned as part of the Users resource. Either the customer or the domain parameter must be provided.}

\item{maxResults}{Maximum number of results to return. Acceptable values are 1 to 500, inclusive. (Default: 100)}
}
\value{
A paginated list of either deleted users or all users in a domain.
}
\description{
Retrieves a paginated list of either deleted users or all users in a domain.
}
\section{References}{

\href{https://developers.google.com/admin-sdk/directory/v1/reference/users/list}{Google API Documentation}
}

\concept{Users}
